[Fix #9928] Fix a false auto-correct for `Layout/EndAlignment`
Fixes #9928.

This PR fixes an infinite loop error and a false auto-correction behavior for `Layout/EndAlignment`
when using operator methods and `EnforcedStyleAlignWith: variable`.
This infinite loop error is due to an offense being registered but not auto-corrected.

`EnforcedStyleAlignWith: variable` enforces alignment to variable,
it should be auto-corrected to be aligned to variable even when operator method is used.